Fix/issue 44 major disputes (#506)
* fixed the issue with major disputes not jailing or deleting the reporter. We now will jail the reporter indefinitely but not deleting as the reporter can become unjailed upon winning the dispute

* switched jailed duration back to max int64

* dealt with an error that was missed in a test

* fixing the same merge issues again for some reason

* linting

* pushing up for help

* added logs to ensure block info is being set in end blocker

* removed fmt.Println() statements

---------

+++ e2e/spinup_test.go
@@ -198,7 +198,8 @@ func TestLearn(t *testing.T) {
 	require.NoError(t, err)
 	err = json.Unmarshal(r, &disputes)
 	require.NoError(t, err)
-	require.Equal(t, disputes.Disputes[0].Metadata.DisputeStatus, 1) // voting
+
+	require.Equal(t, disputes.Disputes[0].Metadata.DisputeStatus, int32(1)) // voting
 	fmt.Println("Disputes: ", string(r))
 	res2, _, err = validatorI.ExecQuery(ctx, "oracle", "get-current-aggregate-report", hex.EncodeToString(qidbz))
 	require.NoError(t, err)
@@ -235,5 +236,5 @@ func TestLearn(t *testing.T) {
 
 	err = json.Unmarshal(r, &disputes)
 	require.NoError(t, err)
-	require.Equal(t, disputes.Disputes[0].Metadata.DisputeStatus, 2) // resolved
+	require.Equal(t, disputes.Disputes[0].Metadata.DisputeStatus, int32(2)) // resolved   DISPUTE_STATUS_RESOLVED
 }
